+++ spec/components/file_upload_spec.rb
@@ -62,6 +62,36 @@ describe "File upload", type: :view do
     assert_select ".govuk-file-upload[accept='image/*']"
   end
 
+  it "renders text input different sized labels" do
+    render_component(
+      label: { text: "What is your email address?" },
+      name: "email-address",
+      heading_size: "xl",
+    )
+
+    assert_select ".govuk-label.govuk-label--xl"
+  end
+
+  it "renders the label wrapped in a heading" do
+    render_component(
+      label: { text: "Where's your head at?" },
+      name: "heading",
+      heading_level: 3,
+    )
+
+    assert_select "h3.govuk-label-wrapper .govuk-label", text: "Where's your head at?"
+  end
+
+  it "only renders a label wrapped in a heading if specified" do
+    render_component(
+      label: { text: "What is your email address?" },
+      name: "email-address",
+      heading_size: "xl",
+    )
+
+    assert_select ".govuk-label-wrapper", false
+  end
+
   context "when a hint is provided" do
     before do
       render_component(