+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 * Standardise search term formatting across GA4 trackers ([PR #3746](https://github.com/alphagov/govuk_publishing_components/pull/3746))
 * Add ZIP file support to attachment component ([PR #3751](https://github.com/alphagov/govuk_publishing_components/pull/3751))
+* Add heading option to file input component ([PR #3755](https://github.com/alphagov/govuk_publishing_components/pull/3755))
 
 ## 36.0.2